Fix some missed size_t updates

+++ ssl/pqueue.c
@@ -141,7 +141,7 @@ pitem *pqueue_next(pitem **item)
     return ret;
 }
 
-int pqueue_size(pqueue *pq)
+size_t pqueue_size(pqueue *pq)
 {
     pitem *item = pq->items;
     int count = 0;
